+++ src/components/Icon/enum/enum.ts
@@ -158,5 +158,8 @@ export enum IconNameEnum {
   crop = 'crop',
   imagePaint = 'image-paint',
   text = 'text',
-  colorPicker = 'color-picker'
+  colorPicker = 'color-picker',
+  connection = 'connection',
+  tableError = 'table-error',
+  rearrange = 'rearrange'
 }